Move syntax types to `syntax.ts`
